# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Physique Classifier API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface PoseLandmark {
  x: number;
  y: number;
  z: number;
  visibility?: number;
}

export type ClassifyRequestGender =
  (typeof ClassifyRequestGender)[keyof typeof ClassifyRequestGender];

export const ClassifyRequestGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface ClassifyRequest {
  /** MediaPipe pose landmarks (33 points) */
  landmarks: PoseLandmark[];
  /** Height in cm (optional) */
  height?: number;
  /** Weight in kg (optional) */
  weight?: number;
  /** Age in years (optional) */
  age?: number;
  gender?: ClassifyRequestGender;
}

export type ClassifyResultPhysiqueType =
  (typeof ClassifyResultPhysiqueType)[keyof typeof ClassifyResultPhysiqueType];

export const ClassifyResultPhysiqueType = {
  athletic: "athletic",
  skinny: "skinny",
  overweight: "overweight",
} as const;

export type ClassifyResultProbabilities = {
  athletic: number;
  skinny: number;
  overweight: number;
};

export interface BodyMetrics {
  shoulderHipRatio: number;
  torsoLegRatio: number;
  symmetryScore: number;
  postureScore: number;
  bmi?: number;
}

export interface ClassifyResult {
  physiqueType: ClassifyResultPhysiqueType;
  /** Confidence score 0-1 */
  confidence: number;
  probabilities: ClassifyResultProbabilities;
  bodyMetrics: BodyMetrics;
  recommendations: string[];
}

export type ExerciseDifficulty =
  (typeof ExerciseDifficulty)[keyof typeof ExerciseDifficulty];

export const ExerciseDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Exercise {
  name: string;
  sets: number;
  reps: string;
  restSeconds: number;
  description: string;
  muscleGroups: string[];
  difficulty: ExerciseDifficulty;
}

export interface DayPlan {
  day: string;
  focus: string;
  exercises: Exercise[];
  durationMinutes: number;
  restDay: boolean;
}

export interface Milestone {
  week: number;
  title: string;
  description: string;
}

export interface WorkoutPlan {
  physiqueType: string;
  goal: string;
  fitnessLevel: string;
  weeklySchedule: DayPlan[];
  nutritionTips: string[];
  progressMilestones: Milestone[];
}

export type AppStatsBreakdown = {
  athletic: number;
  skinny: number;
  overweight: number;
};

export interface AppStats {
  totalClassifications: number;
  breakdown: AppStatsBreakdown;
  averageConfidence: number;
  recentActivity: string[];
}

export type HistoryItemPhysiqueType =
  (typeof HistoryItemPhysiqueType)[keyof typeof HistoryItemPhysiqueType];

export const HistoryItemPhysiqueType = {
  athletic: "athletic",
  skinny: "skinny",
  overweight: "overweight",
} as const;

export interface HistoryItem {
  id: number;
  physiqueType: HistoryItemPhysiqueType;
  confidence: number;
  createdAt: string;
  bodyMetrics: BodyMetrics;
}

export type SaveClassificationRequestPhysiqueType =
  (typeof SaveClassificationRequestPhysiqueType)[keyof typeof SaveClassificationRequestPhysiqueType];

export const SaveClassificationRequestPhysiqueType = {
  athletic: "athletic",
  skinny: "skinny",
  overweight: "overweight",
} as const;

export interface SaveClassificationRequest {
  physiqueType: SaveClassificationRequestPhysiqueType;
  confidence: number;
  bodyMetrics: BodyMetrics;
}

export type GetWorkoutPlanParams = {
  goal?: GetWorkoutPlanGoal;
  fitnessLevel?: GetWorkoutPlanFitnessLevel;
};

export type GetWorkoutPlanGoal =
  (typeof GetWorkoutPlanGoal)[keyof typeof GetWorkoutPlanGoal];

export const GetWorkoutPlanGoal = {
  muscle_gain: "muscle_gain",
  fat_loss: "fat_loss",
  endurance: "endurance",
  flexibility: "flexibility",
  general_fitness: "general_fitness",
} as const;

export type GetWorkoutPlanFitnessLevel =
  (typeof GetWorkoutPlanFitnessLevel)[keyof typeof GetWorkoutPlanFitnessLevel];

export const GetWorkoutPlanFitnessLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;