# lib/api-zod/src/generated/types/appStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Physique Classifier API
 * OpenAPI spec version: 0.1.0
 */
import type { AppStatsBreakdown } from "./appStatsBreakdown";

export interface AppStats {
  totalClassifications: number;
  breakdown: AppStatsBreakdown;
  averageConfidence: number;
  recentActivity: string[];
}