# lib/api-zod/src/generated/types/bodyMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Physique Classifier API
 * OpenAPI spec version: 0.1.0
 */

export interface BodyMetrics {
  shoulderHipRatio: number;
  torsoLegRatio: number;
  symmetryScore: number;
  postureScore: number;
  bmi?: number;
}