# lib/api-zod/src/generated/types/classifyResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Physique Classifier API
 * OpenAPI spec version: 0.1.0
 */
import type { BodyMetrics } from "./bodyMetrics";
import type { ClassifyResultPhysiqueType } from "./classifyResultPhysiqueType";
import type { ClassifyResultProbabilities } from "./classifyResultProbabilities";

export interface ClassifyResult {
  physiqueType: ClassifyResultPhysiqueType;
  /** Confidence score 0-1 */
  confidence: number;
  probabilities: ClassifyResultProbabilities;
  bodyMetrics: BodyMetrics;
  recommendations: string[];
}