# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Physique Classifier API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseDifficulty } from "./exerciseDifficulty";

export interface Exercise {
  name: string;
  sets: number;
  reps: string;
  restSeconds: number;
  description: string;
  muscleGroups: string[];
  difficulty: ExerciseDifficulty;
}