# lib/api-zod/src/generated/types/saveClassificationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fitness Physique Classifier API
 * OpenAPI spec version: 0.1.0
 */
import type { BodyMetrics } from "./bodyMetrics";
import type { SaveClassificationRequestPhysiqueType } from "./saveClassificationRequestPhysiqueType";

export interface SaveClassificationRequest {
  physiqueType: SaveClassificationRequestPhysiqueType;
  confidence: number;
  bodyMetrics: BodyMetrics;
}